cleanup and name changes
